requirements.txt
pocketcasts-api
soco